requirements.txt
anyio==4.4.0
astor==0.8.1
biopython==1.85
biotite==1.2.0
biotraj==1.2.2
certifi==2024.7.4
charset-normalizer==3.4.2
decorator==5.1.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
msgpack==1.1.1
networkx==3.4.2
numpy==2.1.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.9.6.50
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
opt-einsum==3.3.0
packaging==25.0
paddlepaddle-gpu==3.0.0
pandas==2.3.0
pillow==10.4.0
pip==25.1
protobuf==5.27.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scipy==1.15.3
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
